Refactor effect representation
The previous effect of representation of effects as a type-indexed type
family (Type → Type) was causing universe problems. This is an attempt to solve
this. The new representation is that an effect call is a plain type and then
there is an Effect typeclass that provide a function giving the effect return
value.

For end consumer the main change is that events now only have 2 fields, and the
typing prototype are slightly different.

In the process, the concept of an effect associated exception was scrapped
because not every useful. If needed again one can do:
exc := { call : Eff & eff_ret call → False }.

Candidate traces can now contain a trace that is just a list of effect without a
continuing effect, or a return value.